+++ NEWS.md
@@ -1,8 +1,9 @@
 # eatGADS 0.15.2.9000
 
-* bug fix `collapseColumns()` (now supports all column names in look up table)
+* bug fix `collapseColumns()` (now supports `new_value` as column name in look up table)
 * `checkValue()` now returns occurrence count as vector instead of giving a message
 * `removeValLabels()` now can remove only specific `value` - `valLabel` pairs if required 
+* changing value labels inside of `recodeGADS()` via `newValueLabels` deprecated (use `changeValLabels()`instead)
 
 # eatGADS 0.15.2